Добавлена middleware для проверки заголовка хэширования
